Refactor program

+++ test/transform/test_toy.py
@@ -18,7 +18,7 @@ def test_collate():
         t0, t1, t2, out, bsize = collate([x])
         assert t0.dtype == torch.float32
         assert t0.shape == (11, 1, 25)
-        assert t1.dtype == torch.float
+        assert t1.dtype == torch.int64
         assert t2.dtype == torch.bool
         assert len(out) == 1
         assert bsize == 1
